// src/api/generated/models/ListComputeSettingsDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ListComputeSettingsMethodDto } from './ListComputeSettingsMethodDto';

export type ListComputeSettingsDto = {
    computeTypeId: string;
    name: string;
    allowedMethods: Array<ListComputeSettingsMethodDto>;
};